Undo renaming changes for course progress
