+++ assets/blocks/course-progress-block/block.json
