+++ assets/blocks/course-progress-block/index.js
