--- assets/blocks/progress-bar-block/progress-bar-editor.scss
@@ -1,1 +1,0 @@
-@import '../../shared/blocks/progress-bar/progress-bar-settings';
