+++ assets/blocks/single-course-style-editor.scss
@@ -1,2 +1,2 @@
 @import 'course-outline/course-outline-editor';
-@import 'progress-bar-block/progress-bar-editor';
+@import 'course-progress-block/course-progress-editor';
