+++ assets/blocks/single-course-style.scss
@@ -1,2 +1,2 @@
 @import 'course-outline/course-outline';
-@import 'progress-bar-block/progress-bar';
+@import 'course-progress-block/course-progress';
