+++ assets/blocks/single-course.js
@@ -3,7 +3,7 @@
  */
 import registerSenseiBlocks from './register-sensei-blocks';
 import TakeCourseBlock from './take-course-block';
-import ProgressBarBlock from './progress-bar-block';
+import ProgressBarBlock from './course-progress-block';
 import { OutlineBlock, LessonBlock, ModuleBlock } from './course-outline';
 import ConditionalContentBlock from './conditional-content-block';
 import ViewResults from './view-results-block';
